Menu Functionality is complete, options can be selected and entered, entering start puts the game into the dungeon state, entering exit causes the program to exit.
Each state has a default display, which is just a black screen, and a default update, which puts the game back in the menu state.

The game will wait for an input before calling update reguardless of the current state, (this still needs to be updated to the new input method.